+++ xcode/Jasonette/Core/Jason/Jason.m
@@ -157,8 +157,8 @@
             DTLogError (@"Missing $jason property in \n%@\n\n", res);
         }
 
-        VC.original = @{ @"$jason": result };
-        [self drawViewFromJason:VC.original
+        self->VC.original = @{ @"$jason": result };
+        [self drawViewFromJason:self->VC.original
                         asFinal:final];
     }];
 }
@@ -421,8 +421,8 @@
     if (turnon) {
         [JDStatusBarNotification addStyleNamed:@"SBStyle1"
                                        prepare:^JDStatusBarStyle *(JDStatusBarStyle * style) {
-            style.barColor = navigationController.navigationBar.backgroundColor;
-            style.textColor = navigationController.navigationBar.tintColor;
+            style.barColor = self->navigationController.navigationBar.backgroundColor;
+            style.textColor = self->navigationController.navigationBar.tintColor;
             style.animationType = JDStatusBarAnimationTypeFade;
             return style;
         }];
@@ -443,11 +443,11 @@
 - (void)networkLoading:(BOOL)turnon with:(NSDictionary *)options; {
     dispatch_async (dispatch_get_main_queue (), ^{
         if (turnon && (options == nil || (options != nil && options[@"loading"] && [options[@"loading"] boolValue]))) {
-            MBProgressHUD * hud = [MBProgressHUD showHUDAddedTo:VC.view animated:true];
+            MBProgressHUD * hud = [MBProgressHUD showHUDAddedTo:self->VC.view animated:true];
             hud.animationType = MBProgressHUDAnimationFade;
             hud.userInteractionEnabled = NO;
         } else if (!turnon) {
-            [MBProgressHUD hideHUDForView:VC.view animated:true];
+            [MBProgressHUD hideHUDForView:self->VC.view animated:true];
         }
     });
 }
@@ -620,7 +620,7 @@
                                                                 image:nil
                                                      highlightedImage:nil
                                                                action:^(REMenuItem * item) {
-                                                                   [menu_component close];
+                                                                   [self->menu_component close];
 
                                                                    if (item_action) {
                                                                    [memory set_stack:item_action];
@@ -1276,16 +1276,17 @@
 #pragma message "Start Request in Include"
                 [manager   GET:url
                     parameters:parameters
+                       headers:nil
                       progress:^(NSProgress * _Nonnull downloadProgress) { }
                        success:^(NSURLSessionDataTask * _Nonnull task, id _Nullable responseObject)
                 {
-                    VC.requires[url] = responseObject;
+                    self->VC.requires[url] = responseObject;
                     dispatch_group_leave (requireGroup);
                 }
                        failure:^(NSURLSessionDataTask * _Nullable task, NSError * _Nonnull error)
                 {
                     DTLogWarning (@"Error Fetching JSON From Url %@ %@", url, error);
-                    VC.requires[url] = @{};
+                    self->VC.requires[url] = @{};
                     dispatch_group_leave (requireGroup);
                 }];
             }
@@ -1404,6 +1405,7 @@
 #pragma message "Start Request in Require"
                 [manager   GET:url
                     parameters:parameters
+                       headers:nil
                       progress:^(NSProgress * _Nonnull downloadProgress) { }
                        success:^(NSURLSessionDataTask * _Nonnull task, id _Nullable responseObject)
                 {
@@ -1447,7 +1449,7 @@
         // require could take a long time to finish, so we make sure at this point
         // we are looking at the same URL we began with
         [self success:dict withOriginalUrl:origin_url];
-        [MBProgressHUD hideHUDForView:VC.view animated:true];
+        [MBProgressHUD hideHUDForView:self->VC.view animated:true];
     });
 }
 
@@ -1844,8 +1846,8 @@
     VC.rendered = nil;
 
     dispatch_async (dispatch_get_main_queue (), ^{
-        [VC viewDidLoad];
-        [VC viewWillAppear:NO];
+        [self->VC viewDidLoad];
+        [self->VC viewWillAppear:NO];
     });
 }
 
@@ -2225,6 +2227,7 @@
 
             [manager   GET:self->VC.url
                 parameters:parameters
+                   headers:nil
                   progress:^(NSProgress * _Nonnull downloadProgress) { }
                    success:^(NSURLSessionDataTask * _Nonnull task, id _Nullable responseObject) {
                        // Ignore if the url is different
@@ -2242,7 +2245,7 @@
                                         self->VC.contentLoaded = NO;
 
                                         self->VC.original = @{ @"$jason": res[@"$jason"] };
-                                        [self drawViewFromJason:VC.original
+                                        [self drawViewFromJason:self->VC.original
                                                         asFinal:YES];
                                     });
                        }];
@@ -2426,7 +2429,7 @@
 
 - (void)drawAdvancedBackground:(NSDictionary *)bg {
     dispatch_async (dispatch_get_main_queue (), ^{
-        [self drawAdvancedBackground:bg forVC:VC];
+        [self drawAdvancedBackground:bg forVC:self->VC];
     });
 }
 
@@ -2523,30 +2526,30 @@
 
             CGRect rect = CGRectMake (x, y, width, height);
 
-            if (@available(iOS 11, *)) {
-                // Take in consideration safe areas available in iOS 11
-                y = -vc.view.safeAreaInsets.top;
-                height = [UIScreen mainScreen].bounds.size.height +
-                    vc.view.safeAreaInsets.top +
-                    vc.view.safeAreaInsets.bottom;
-
-
-                x = -vc.view.safeAreaInsets.left;
-                width = [UIScreen mainScreen].bounds.size.width +
-                    vc.view.safeAreaInsets.left +
-                    vc.view.safeAreaInsets.right;
-
-                if (!tabController.tabBar.hidden) {
-                    height = height - tabController.tabBar.frame.size.height;
-                }
-
-                if (vc.composeBarView) {
-                    // footer.input exists
-                    height = height - vc.composeBarView.frame.size.height;
-                }
-
-                rect = CGRectMake (x, y, width, height);
-            }
+//            if (@available(iOS 11, *)) {
+//                // Take in consideration safe areas available in iOS 11
+//                y = -vc.view.safeAreaInsets.top;
+//                height = [UIScreen mainScreen].bounds.size.height +
+//                    vc.view.safeAreaInsets.top +
+//                    vc.view.safeAreaInsets.bottom;
+//
+//
+//                x = -vc.view.safeAreaInsets.left;
+//                width = [UIScreen mainScreen].bounds.size.width +
+//                    vc.view.safeAreaInsets.left +
+//                    vc.view.safeAreaInsets.right;
+//
+//                if (!tabController.tabBar.hidden) {
+//                    height = height - tabController.tabBar.frame.size.height;
+//                }
+//
+//                if (vc.composeBarView) {
+//                    // footer.input exists
+//                    height = height - vc.composeBarView.frame.size.height;
+//                }
+//
+//                rect = CGRectMake (x, y, width, height);
+//            }
 
             vc.background.frame = rect;
 
@@ -2569,6 +2572,8 @@
             [progressView setFrame:CGRectMake (0, vc.background.frame.origin.y + navHeight + 20, vc.background.frame.size.width, progressView.frame.size.height)];
             [progressView setAutoresizingMask:UIViewAutoresizingFlexibleWidth | UIViewAutoresizingFlexibleTopMargin];
 
+#pragma message "Webview Styles"
+
             if (bg[@"style"]) {
                 if (bg[@"style"][@"background"]) {
                     vc.background.backgroundColor = [JasonHelper colorwithHexString:bg[@"style"][@"background"] alpha:1.0];
@@ -2586,17 +2591,18 @@
 }
 
 - (void)buildCamera:(NSDictionary *)options forVC:(JasonViewController *)vc {
+    DTLogDebug (@"Building Camera");
+
+
     NSError * error = nil;
     // Find back/front camera
     // based on options
     NSArray * devices = [AVCaptureDevice devicesWithMediaType:AVMediaTypeVideo];
     AVCaptureDevice * device = [AVCaptureDevice defaultDeviceWithMediaType:AVMediaTypeVideo];
-    AVCaptureDevicePosition position;
+    AVCaptureDevicePosition position = AVCaptureDevicePositionFront;
 
     if (options[@"device"] && [options[@"device"] isEqualToString:@"back"]) {
         position = AVCaptureDevicePositionBack;
-    } else {
-        position = AVCaptureDevicePositionFront;
     }
 
     for (AVCaptureDevice * d in devices) {
@@ -2610,9 +2616,33 @@
 
     // Add input to the session
     AVCaptureDeviceInput * input = [AVCaptureDeviceInput deviceInputWithDevice:device error:&error];
+
+    if (error) {
+        DTLogWarning (@"$vision: %@", error);
+    }
+
+    if (![self.avCaptureSession canAddInput:input]) {
+        DTLogWarning (@"$vision: No camera found. Are you using a simulator?");
+        DTLogDebug (@"Loading error.json");
+
+        [self loadViewByFile:@"error.json" asFinal:YES];
+
+        [self call:@{ @"type": @"$util.alert",
+                      @"options": @{
+                          @"description": @"No camera found."
+        } }];
+        return;
+    }
+
     [self.avCaptureSession addInput:input];
 
     AVCaptureMetadataOutput * output = [[AVCaptureMetadataOutput alloc] init];
+
+    if (![self.avCaptureSession canAddOutput:output]) {
+        DTLogWarning (@"$vision: Can't add output");
+        return;
+    }
+
     [self.avCaptureSession addOutput:output];
 
     // Listen for different types of barcode detection
@@ -2636,7 +2666,7 @@
 
 - (void)drawBackground:(NSString *)bg {
     dispatch_async (dispatch_get_main_queue (), ^{
-        [self drawBackground:bg forVC:VC];
+        [self drawBackground:bg forVC:self->VC];
     });
 }
 
@@ -3366,17 +3396,17 @@
             // If the tabs are empty AND the view hasn't been rendered yet, then wait until it finishes rendering
 
             if (!t && !v.rendered) {
-                if (previous_footer && previous_footer[@"tabs"]) {
+                if (self->previous_footer && self->previous_footer[@"tabs"]) {
                     // don't touch yet until the view finalizes
                 } else {
-                    tabController.tabBar.hidden = YES;
+                    self->tabController.tabBar.hidden = YES;
                 }
 
                 return;
             }
         }
 
-        if (previous_footer && previous_footer[@"tabs"]) {
+        if (self->previous_footer && self->previous_footer[@"tabs"]) {
             // if previous footer tab was not null, we diff the tabs to determine whether to re-render
             if (v.old_footer && v.old_footer[@"tabs"] && [[v.old_footer[@"tabs"] description] isEqualToString:[t description]]) {
                 return;
@@ -3391,18 +3421,18 @@
 
         v.old_footer[@"tabs"] = t;
 
-        if (!previous_footer) {
-            previous_footer = [[NSMutableDictionary alloc] init];
+        if (!self->previous_footer) {
+            self->previous_footer = [[NSMutableDictionary alloc] init];
         }
 
-        previous_footer[@"tabs"] = t;
+        self->previous_footer[@"tabs"] = t;
 
         if (!t || [t isKindOfClass:[NSNull class]]) {
-            tabController.tabBar.hidden = YES;
-            tabController.viewControllers = @[navigationController]; // remove all tab bar items if there's no "items"
+            self->tabController.tabBar.hidden = YES;
+            self->tabController.viewControllers = @[self->navigationController]; // remove all tab bar items if there's no "items"
             return;
         } else {
-            tabController.tabBar.hidden = NO;
+            self->tabController.tabBar.hidden = NO;
         }
 
         NSArray * tabs = t[@"items"];
@@ -3411,7 +3441,7 @@
         if (style) {
             if (style[@"color"]) {
                 UIColor * c = [JasonHelper colorwithHexString:style[@"color"] alpha:1.0];
-                [tabController.tabBar setTintColor:c];
+                [self->tabController.tabBar setTintColor:c];
                 [[UITabBarItem appearance] setTitleTextAttributes:@{ NSForegroundColorAttributeName: c }
                                                          forState:UIControlStateSelected];
             }
@@ -3424,11 +3454,11 @@
             }
 
             if (style[@"background"]) {
-                [tabController.tabBar setClipsToBounds:YES];
-                tabController.tabBar.shadowImage = [[UIImage alloc] init];
-                tabController.tabBar.translucent = NO;
-                tabController.tabBar.backgroundColor = [JasonHelper colorwithHexString:style[@"background"] alpha:1.0];
-                [tabController.tabBar setBarTintColor:[JasonHelper colorwithHexString:style[@"background"] alpha:1.0]];
+                [self->tabController.tabBar setClipsToBounds:YES];
+                self->tabController.tabBar.shadowImage = [[UIImage alloc] init];
+                self->tabController.tabBar.translucent = NO;
+                self->tabController.tabBar.backgroundColor = [JasonHelper colorwithHexString:style[@"background"] alpha:1.0];
+                [self->tabController.tabBar setBarTintColor:[JasonHelper colorwithHexString:style[@"background"] alpha:1.0]];
             }
 
             [[UITabBar appearance] setTranslucent:NO];
@@ -3450,7 +3480,7 @@
             // that was initialized with
             // In this case, initialize all tabs
             // Start from index 1 because the first one should already be instantiated via modal href
-            if (tabController.viewControllers.count != maxTabCount) {
+            if (self->tabController.viewControllers.count != maxTabCount) {
                 firstTime = YES;
                 tabs_array = [[NSMutableArray alloc] init];
             } else {
@@ -3462,18 +3492,18 @@
             // because when we transition from view A with 5 tabs to view B with no tab, Jasonette gets rid of all the navigation controllers
             // so when we come back from view B, the tab bar that contains view A will only have one item, and it will say the index is 0, which is incorrect.
             // To avoid this situation, we need to be more precise and decide on the index based on the view's URL instead.
-            NSUInteger indexOfTab = [tabController.viewControllers indexOfObject:navigationController];
-            tabNeedsRefresh = YES;
+            NSUInteger indexOfTab = [self->tabController.viewControllers indexOfObject:self->navigationController];
+            self->tabNeedsRefresh = YES;
 
             for (int i = 0; i < maxTabCount; i++) {
                 NSDictionary * tab = tabs[i];
 
-                if (tab[@"url"] && [VC.url isEqualToString:tab[@"url"]]) {
+                if (tab[@"url"] && [self->VC.url isEqualToString:tab[@"url"]]) {
                     indexOfTab = i;
-                    tabNeedsRefresh = NO;
-                } else if (tab[@"href"] && tab[@"href"][@"url"] && [VC.url isEqualToString:tab[@"href"][@"url"]]) {
+                    self->tabNeedsRefresh = NO;
+                } else if (tab[@"href"] && tab[@"href"][@"url"] && [self->VC.url isEqualToString:tab[@"href"][@"url"]]) {
                     indexOfTab = i;
-                    tabNeedsRefresh = NO;
+                    self->tabNeedsRefresh = NO;
                 }
             }
 
@@ -3508,15 +3538,15 @@
                     // no need to create a new VC, etc. because it's already been instantiated
                         tabFound = YES;
                     // if the tab URL is same as the currently visible VC's url
-                        VC.tabNeedsRefresh = YES;
-                        [tabs_array addObject:navigationController];
+                        self->VC.tabNeedsRefresh = YES;
+                        [tabs_array addObject:self->navigationController];
                     } else {
                     // for all other tabs, create a new VC and instantiate them, and add them to the tabs array
                         JasonViewController * vc = [[JasonViewController alloc] init];
                         vc.url = url;
 
-                        if (tabNeedsRefresh) {
-                            vc.tabNeedsRefresh = tabNeedsRefresh;
+                        if (self->tabNeedsRefresh) {
+                            vc.tabNeedsRefresh = self->tabNeedsRefresh;
                         }
 
                         vc.options = [self filloutTemplate:options withData:[self variables]];
@@ -3533,13 +3563,13 @@
                         v.tabNeedsRefresh = YES;
                         tabFound = YES;
                     } else {
-                        UINavigationController * nav = tabController.viewControllers[i];
+                        UINavigationController * nav = self->tabController.viewControllers[i];
                         JasonViewController * vc = [[nav viewControllers] firstObject];
 
                         vc.url = url;
 
-                        if (tabNeedsRefresh) {
-                            vc.tabNeedsRefresh = tabNeedsRefresh;
+                        if (self->tabNeedsRefresh) {
+                            vc.tabNeedsRefresh = self->tabNeedsRefresh;
                         }
 
                         vc.options = [self filloutTemplate:options withData:[self variables]];
@@ -3550,17 +3580,17 @@
             }
 
             if (firstTime) {
-                tabController.viewControllers = tabs_array;
+                self->tabController.viewControllers = tabs_array;
             }
 
             for (int i = 0; i < maxTabCount; i++) {
                 NSDictionary * tab = tabs[i];
-                [self setTabBarItem:[tabController.tabBar.items objectAtIndex:i] withTab:tab];
+                [self setTabBarItem:[self->tabController.tabBar.items objectAtIndex:i] withTab:tab];
             }
 
-            tabController.tabBar.hidden = NO;
+            self->tabController.tabBar.hidden = NO;
         } else {
-            tabController.tabBar.hidden = YES;
+            self->tabController.tabBar.hidden = YES;
         }
     });
 }
@@ -3898,9 +3928,9 @@
     *******************************/
     dispatch_async (dispatch_get_main_queue (), ^{
         // Dismiss searchbar before transitioning.
-        if (VC.searchController) {
-            if (VC.searchController.isActive) {
-                [VC.searchController setActive:NO];
+        if (self->VC.searchController) {
+            if (self->VC.searchController.isActive) {
+                [self->VC.searchController setActive:NO];
             }
         }
 
@@ -3909,7 +3939,7 @@
         NSString * fresh = href[@"fresh"];
         JasonMemory * memory = [JasonMemory client];
         memory.executing = NO;
-        queue = [@[] mutableCopy];
+        self->queue = [@[] mutableCopy];
 
         if ([transition isEqualToString:@"root"]) {
             [self start:nil];
@@ -3948,10 +3978,10 @@
                 DTLogDebug (@"Presenting as modal transition");
                 UINavigationController * newNav = [[UINavigationController alloc]initWithRootViewController:vc];
                 [newNav setNavigationBarHidden:YES animated:NO];
-                [navigationController presentViewController:newNav animated:YES completion:^{ }];
+                [self->navigationController presentViewController:newNav animated:YES completion:^{ }];
             } else {
                 DTLogDebug (@"Presenting as push transition");
-                [navigationController presentViewController:vc animated:YES completion:^{ }];
+                [self->navigationController presentViewController:vc animated:YES completion:^{ }];
             }
         } else if ([view.lowercaseString isEqualToString:@"app"] || [view.lowercaseString isEqualToString:@"external"]) {
             /****************************************************************************
@@ -3961,6 +3991,8 @@
             ****************************************************************************/
             NSString * url = href[@"url"];
 
+#pragma message "$href action"
+
             DTLogDebug (@"Opening External URL %@", url);
 
             if (memory._register && memory._register.count > 0) {
@@ -3969,7 +4001,41 @@
             }
 
             if (url) {
-                [[UIApplication sharedApplication] openURL:[NSURL URLWithString:url]];
+                // See https://developer.apple.com/library/archive/featuredarticles/iPhoneURLScheme_Reference/Introduction/Introduction.html#//apple_ref/doc/uid/TP40007899
+                BOOL systemSchemes = ([url hasPrefix:@"tel:"] ||
+                                      [url hasPrefix:@"sms:"] ||
+                                      [url hasPrefix:@"mailto:"] ||
+                                      [url hasPrefix:@"facetime:"] ||
+                                      [url hasPrefix:@"facetime-audio:"]);
+
+                if (systemSchemes) {
+                    #if TARGET_IPHONE_SIMULATOR
+                    DTLogWarning (@"Calling tel:, sms:, mailto:, facetime:, facetime-audio: urls in simulator do not work. Test them in a device.");
+                    [self call:@{ @"type": @"$util.alert",
+                                  @"options": @{
+                                      @"description": @"This action should be tested in a real device"
+                    } }];
+                    #endif
+                }
+
+                // Secure the url as per apple guidelines
+                // in https://developer.apple.com/documentation/uikit/inter-process_communication/allowing_apps_and_websites_to_link_to_your_content/defining_a_custom_url_scheme_for_your_app
+
+                NSMutableCharacterSet * chars = [[NSCharacterSet URLHostAllowedCharacterSet] mutableCopy];
+                [chars formUnionWithCharacterSet:[NSCharacterSet URLPathAllowedCharacterSet]];
+                [chars formUnionWithCharacterSet:[NSCharacterSet URLQueryAllowedCharacterSet]];
+                [chars formUnionWithCharacterSet:[NSCharacterSet URLFragmentAllowedCharacterSet]];
+                [chars formUnionWithCharacterSet:[NSCharacterSet URLUserAllowedCharacterSet]];
+                [chars formUnionWithCharacterSet:[NSCharacterSet URLPasswordAllowedCharacterSet]];
+
+                NSString * encodedUrl = [[url stringByRemovingPercentEncoding]
+                                         stringByAddingPercentEncodingWithAllowedCharacters:chars];
+
+                NSURL * destination = [NSURL URLWithString:encodedUrl];
+
+                DTLogDebug (@"openURL: %@", encodedUrl);
+
+                [[UIApplication sharedApplication] openURL:destination];
             } else {
                 DTLogWarning (@"Invalid Url");
             }
@@ -4101,8 +4167,8 @@
 
                     vc.extendedLayoutIncludesOpaqueBars = YES;
 
-                    if (tabController.tabBar.hidden) {
-                        tabController.tabBar.hidden = YES;
+                    if (self->tabController.tabBar.hidden) {
+                        self->tabController.tabBar.hidden = YES;
                     } else {
                         if ([transition isEqualToString:@"fullscreen"]) {
                             vc.hidesBottomBarWhenPushed = YES;
@@ -4111,7 +4177,7 @@
                         }
                     }
 
-                    [navigationController pushViewController:vc animated:YES];
+                    [self->navigationController pushViewController:vc animated:YES];
                 }
             } else {
                 /****************************************************************************
@@ -4243,10 +4309,10 @@
                     }
 
                     [self unlock];
-                    [navigationController presentViewController:nav
-                                                       animated:YES
-                                                     completion:^{
-                                                     }];
+                    [self->navigationController presentViewController:nav
+                                                             animated:YES
+                                                           completion:^{
+                                                           }];
                     CFRunLoopWakeUp (CFRunLoopGetCurrent ());
                 }
                 // Option 2. Push transition
@@ -4258,7 +4324,7 @@
                     }
 
                     [self unlock];
-                    [navigationController pushViewController:vc animated:YES];
+                    [self->navigationController pushViewController:vc animated:YES];
                 }
             }
         }
@@ -4693,21 +4759,21 @@
     // Experimental: Store cache content for offline
     DTLogInfo (@"Loading View From Cache");
 
-    if (VC.original && VC.rendered && VC.original[@"$jason"][@"head"][@"offline"]) {
+    if (self->VC.original && self->VC.rendered && self->VC.original[@"$jason"][@"head"][@"offline"]) {
         DTLogInfo (@"Offline Mode Activated");
 
         if (![[VC.rendered description] containsString:@"{{"] && ![[self.options description] containsString:@"}}"]) {
             dispatch_async (dispatch_get_global_queue (DISPATCH_QUEUE_PRIORITY_HIGH, 0), ^{
-                NSString * normalized_url = [JasonHelper normalized_url:VC.url forOptions:VC.options];
+                NSString * normalized_url = [JasonHelper normalized_url:self->VC.url forOptions:self->VC.options];
                 normalized_url = [normalized_url stringByReplacingOccurrencesOfString:@"/" withString:@"_"];
                 NSArray * paths = NSSearchPathForDirectoriesInDomains (NSDocumentDirectory, NSUserDomainMask, YES);
                 NSString * documentsDirectory = [paths objectAtIndex:0];
                 NSString * path = [documentsDirectory stringByAppendingPathComponent:normalized_url];
 
-                NSMutableDictionary * to_store = [VC.original mutableCopy];
+                NSMutableDictionary * to_store = [self->VC.original mutableCopy];
 
                 if (to_store[@"$jason"]) {
-                    to_store[@"$jason"][@"body"] = VC.rendered;
+                    to_store[@"$jason"][@"body"] = self->VC.rendered;
                 }
 
                 NSData * data = [NSKeyedArchiver archivedDataWithRootObject:to_store];
@@ -4718,7 +4784,7 @@
     }
 
     // if not offline, delete the file associated with the url
-    NSString * normalized_url = [JasonHelper normalized_url:VC.url forOptions:VC.options];
+    NSString * normalized_url = [JasonHelper normalized_url:self->VC.url forOptions:self->VC.options];
     normalized_url = [normalized_url stringByReplacingOccurrencesOfString:@"/" withString:@"_"];
     NSArray * paths = NSSearchPathForDirectoriesInDomains (NSDocumentDirectory, NSUserDomainMask, YES);
     NSString * documentsDirectory = [paths objectAtIndex:0];